requirements.txt
cffi==1.5.2
click==6.3
cryptography==1.3.1
dj-database-url==0.4.0
Django==1.9.4
django-pgcrypto==1.3.0
django-test-without-migrations==0.4
djangorestframework==3.3.3
enum34==1.1.2
idna==2.1
ipaddress==1.0.16
model-mommy==1.2.6
ordereddict==1.1
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
python-dotenv==0.4.0
six==1.10.0
